backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
alembic
asyncpg
celery
flower
redis
playwright
beautifulsoup4
scrapy
anthropic
python-jose[cryptography]
passlib[bcrypt]
bcrypt
pydantic-settings
email-validator
structlog
sentry-sdk
slowapi
httpx
psycopg2-binary
pgvector
itsdangerous
feedparser
